Added the 'Serial Multiplexer Moisture' device to r2Project. It required some refactoring, by allowing custom parameters to be sent for the get-value request (since the multiplexer can have several values it has to work differently). Also: Some bug fixes and cleanup

+++ Arduino/r2Moist/r2Moist.cpp
@@ -73,7 +73,7 @@ R2Moist::R2Moist(R2Multiplexer* multiplexer,
 				 DEVICE_PORT analogPort, 
 				 const uint8_t (&controlPorts)[SENSOR_ROD_COUNT],
 				 uint8_t* sensorChannels,
-				 size_t sensorPairCount) {
+				 uint8_t sensorPairCount) {
 
 	_multiplexer = multiplexer;
 	_analogPort = analogPort;
